pyproject.toml
[project]
name = "local-livekit-plugins"
version = "0.1.0"
description = "Local STT and TTS plugins for LiveKit Agents - no cloud APIs required"
readme = "README.md"
license = {text = "MIT"}
authors = [
    {name = "CoreWorxLab", email = "cmac@coreworxlab.com"}
]
requires-python = ">=3.10"
keywords = [
    "livekit",
    "voice",
    "speech-to-text",
    "text-to-speech",
    "whisper",
    "piper",
    "local",
    "self-hosted"
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Multimedia :: Sound/Audio :: Speech",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]
dependencies = [
    "livekit-agents[silero,turn-detector]~=1.2",
    "livekit-plugins-openai~=1.0",
    "faster-whisper>=1.0.0",
    "piper-tts>=1.2.0",
    "numpy>=1.24.0",
    "python-dotenv>=1.0.0",
    "torch>=2.0.0",  # Provides CUDA 12 libraries for faster-whisper
    "librosa>=0.11.0",
    "httpx>=0.28.1",
    "websockets>=15.0.1",
]

[project.optional-dependencies]
cloud = [
    "livekit-plugins-deepgram~=1.0",
    "livekit-plugins-cartesia~=1.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/local_livekit_plugins"]

[dependency-groups]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "mypy>=1.0.0",
    "ruff>=0.1.0",
]

[tool.uv]
default-groups = ["dev"]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W"]

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
